reorganise folder structure, update triage, add simplify rule for allow trailing zeros
