+++ src/__tests__/latex-equal.spec.ts
@@ -13,9 +13,14 @@ const args = minimist(
 
 const cwd = resolve(__dirname, "../..");
 
-const fixtures = sync(args.t || "src/fixtures/latex-equal/**.ts", {
-  cwd,
-});
+const fixtures = sync(
+  args.t ||
+    ("src/fixtures/latex-equal/literal/**.ts" &&
+      "src/fixtures/latex-equal/symbolic/**.ts"),
+  {
+    cwd,
+  }
+);
 
 let testData = [];
