+++ src/fixtures/latex-equal/literal/PD-1031.ts
@@ -1,8 +1,8 @@
-import { Triage } from "../triage";
+import { Triage } from "../../triage";
 
 export default {
   mode: "literal",
- // skip: true,
+  skip: true,
   tests: [{ target: "1000{mL}", eq: ["1000ml"] }],
   triage: Triage.MEASUREMENT_UNITS,
 };
