+++ src/fixtures/latex-equal/literal/PD-283.ts
@@ -1,8 +1,6 @@
-import { Triage } from "../triage";
-
 export default {
   mode: "literal",
-  //skip: true,
+  skip: true,
   tests: [
     {
       opts: {
