+++ src/fixtures/latex-equal/literal/old-literal.ts
@@ -6,7 +6,7 @@
  */
 export default {
   mode: "literal",
-  skip: false,
+  skip: true,
   tests: [
     { target: "1,000", eq: ["1000"] },
     {
