+++ src/fixtures/latex-equal/symbolic/7119.ts
@@ -1,6 +1,6 @@
 export default {
   mode: "symbolic",
-  skip: true,
+  skip: false,
   // all passed
   tests: [
     {
