+++ src/fixtures/latex-equal/symbolic/bad-input.ts
@@ -1,11 +1,9 @@
-import { Triage } from "../triage";
 export default {
-  skip: true,
+  skip: false,
   mode: "symbolic",
   tests: [
     // all passed
     {
-      triage: [Triage.BAD_USER_INPUT],
       target: "1 + 1",
       ne: [
         "abcdefghijklmnopqrstuvwxyz 1 2 abcdeghijk ab cd ef bq arub oin tsn arstoin a8 asin aosin cvbkyu isn",
