+++ src/fixtures/latex-equal/symbolic/fractions-symbolic.ts
@@ -1,8 +1,8 @@
-import { Triage } from "../triage";
+import { Triage } from "../../triage";
 
 export default {
   mode: "symbolic",
-  skip: true,
+  //skip: true,
   tests: [
     {
       target: "\\frac{x}{12}\\times7\\text{dollars}",
